deps: minimize deps
