+++ packages/@roots/merged-manifest-webpack-plugin/package.json
@@ -69,7 +69,6 @@
   "dependencies": {
     "fs-extra": "^10.0.0",
     "helpful-decorators": "^2.1.0",
-    "prettier": "^2.4.1",
     "tslib": "^2.3.1"
   },
   "stableVersion": "4.6.0"
