+++ packages/@roots/wordpress-dependencies-webpack-plugin/src/index.ts
@@ -9,4 +9,4 @@
  * @packageDocumentation
  */
 
-export {WordPressDependenciesWebpackPlugin} from './WordPressDependenciesWebpackPlugin'
+export {WordPressDependenciesWebpackPlugin} from './plugin'
