+++ packages/@roots/wordpress-dependencies-webpack-plugin/tsconfig.json
@@ -5,13 +5,9 @@
     "rootDir": "./src",
     "outDir": "./lib/cjs",
     "declarationDir": "./types",
-    "types": [
-      "node"
-    ]
+    "types": ["node"]
   },
-  "include": [
-    "src"
-  ],
+  "include": ["src"],
   "references": [
     {
       "path": "./../bud-framework/tsconfig.json"
