+++ packages/@roots/wordpress-externals-webpack-plugin/src/index.ts
@@ -10,25 +10,4 @@
  * @packageDocumentation
  */
 
-import {Compiler, ExternalsPlugin} from 'webpack'
-
-import {externals} from './externals'
-
-/**
- * @public
- */
-export class WordPressExternals {
-  public name = 'WordPressExternalsWebpackPlugin'
-
-  public stage = Infinity
-
-  public externals: ExternalsPlugin
-
-  public constructor() {
-    this.externals = new ExternalsPlugin('window', externals)
-  }
-
-  public apply(compiler: Compiler): void {
-    this.externals.apply(compiler)
-  }
-}
+export {WordPressExternals} from './plugin'
